Feat : api 호출 컴포넌트 error-boundary를 통해 예외처리
