+++ src/pages/home/index.tsx
@@ -1,5 +1,3 @@
-import { Suspense } from "react";
-import { ApiErrorBoundary } from "../common/components/ApiErrorBoundary";
 import { authStatusType, Layout } from "../common/components/Layout";
 import { HomeMain } from "./components/HomeMain";
 import { HomeSideBar } from "./components/HomeSideBar";
@@ -7,14 +5,10 @@ import { HomeSideBar } from "./components/HomeSideBar";
 export const Home = () => {
   return (
     <Layout authStatus={authStatusType.NEED_LOGIN}>
-      <ApiErrorBoundary>
-        <Suspense>
-          <div className="w-full flex gap-6">
-            <HomeSideBar />
-            <HomeMain />
-          </div>
-        </Suspense>
-      </ApiErrorBoundary>
+      <div className="w-full flex gap-6">
+        <HomeSideBar />
+        <HomeMain />
+      </div>
     </Layout>
   );
 };
