+++ src/pages/purchaseHistory/index.tsx
@@ -1,3 +1,5 @@
+import { Suspense } from "react";
+import { ApiErrorBoundary } from "../common/components/ApiErrorBoundary";
 import { authStatusType, Layout } from "../common/components/Layout";
 import { PurchaseList } from "./components/PurchaseList";
 
@@ -6,7 +8,11 @@ export const PurchaseHistory = () => {
     <Layout authStatus={authStatusType.NEED_BUYER}>
       <div className="flex flex-col gap-5">
         <p className="text-3xl font-bold text-title">주문 내역</p>
-        <PurchaseList />
+        <ApiErrorBoundary>
+          <Suspense fallback={<div>Loading...</div>}>
+            <PurchaseList />
+          </Suspense>
+        </ApiErrorBoundary>
       </div>
     </Layout>
   );
